requirements.txt
# Weibo Risk Perception System - Python Dependencies
# Core packages only - most should already be installed

# Core data processing
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# Chinese NLP
jieba>=0.42.1
snownlp>=0.12.3

# Deep learning
torch>=2.0.0
transformers>=4.30.0

# Machine learning
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Network analysis
networkx>=2.8

# Data collection
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Visualization
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
matplotlib>=3.5.0

# Web framework
flask>=2.2.0

# Configuration
pyyaml>=6.0

# Logging
loguru>=0.7.0

# Progress bar
tqdm>=4.65.0

# Other utilities
python-dateutil>=2.8.2
jinja2>=3.1.0

